long max and min values

+++ Jakarta/src/main/jakarta/Main.java
@@ -3,8 +3,8 @@ package main.jakarta;
 public class Main {
     public static void main(String[] args) {
         int myValue = 10000;
-        int myMinIntValue = Integer.MIN_VALUE;
-        int myMaxIntValue = Integer.MAX_VALUE;
+        long myMinIntValue = Long.MIN_VALUE;
+        long myMaxIntValue = Long.MAX_VALUE;
         System.out.println("myMaxIntValue = " + myMaxIntValue);
         System.out.println("myMinIntValue = " + myMinIntValue);
     }
